requirements.txt
python-telegram-bot==20.3
yt-dlp==2024.3.10